feat: Implement AutoMapper and Refactor Event Creation Process
- Added AutoMapper for mapping EventDto to Event entity.
- Refactored CreateEventUseCase to use AutoMapper for converting DTO to Entity.
- Adjusted EventController to handle the returned Event entity
- Added MappingProfile in the Application layer for DTO to Entity mappings."
