+++ EventMS.Application/UseCases/CreateEventUseCase.cs
@@ -1,4 +1,6 @@
-﻿using EventMS.Application.Port;
+﻿using AutoMapper;
+using EventMS.Application.DTOs;
+using EventMS.Application.Port;
 using EventMS.Domain.Entities;
 using EventMS.Domain.Interfaces;
 using System;
@@ -12,14 +14,18 @@ namespace EventMS.Application.UseCases
     public class CreateEventUseCase : ICreateEventUseCase
     {
         private readonly IEventRepository _eventRepository;
+        private readonly IMapper _mapper;
 
-        public CreateEventUseCase(IEventRepository eventRepository)
+        public CreateEventUseCase(IEventRepository eventRepository, IMapper mapper)
         {
             _eventRepository = eventRepository;
+            _mapper = mapper;
         }
 
-        public void Execute(Event newEvent)
+        public Event Execute(EventDto newEventDto)
         {
+            var newEvent = _mapper.Map<Event>(newEventDto);
+
             if (string.IsNullOrWhiteSpace(newEvent.Title) || newEvent.Date == default || newEvent.Time == default || string.IsNullOrWhiteSpace(newEvent.Location))
             {
                 throw new ArgumentException("Missing mandatory fields: title, date, time, location.");
@@ -31,6 +37,7 @@ namespace EventMS.Application.UseCases
             }
 
             _eventRepository.AddEvent(newEvent);
+            return newEvent;
         }
     }
 }
